chore: CON-1341 Inline remaining tECDSA tests (#619)
I will fix "ecdsa" naming in a follow up

---------
