+++ rs/tests/Cargo.toml
@@ -200,38 +200,6 @@ name = "ic-systest-rotate-ecdsa-idkg-key"
 path = "consensus/orchestrator/rotate_ecdsa_idkg_key_test.rs"
 
 [[bin]]
-name = "ic-systest-tecdsa-key-rotation"
-path = "consensus/tecdsa/tecdsa_key_rotation_test.rs"
-
-[[bin]]
-name = "ic-systest-tecdsa-signature-life-cycle"
-path = "consensus/tecdsa/tecdsa_signature_life_cycle_test.rs"
-
-[[bin]]
-name = "ic-systest-tecdsa-signature-fails-without-cycles"
-path = "consensus/tecdsa/tecdsa_signature_fails_without_cycles_test.rs"
-
-[[bin]]
-name = "ic-systest-tecdsa-signature-from-nns-without-cycles"
-path = "consensus/tecdsa/tecdsa_signature_from_nns_without_cycles_test.rs"
-
-[[bin]]
-name = "ic-systest-tecdsa-signature-from-other-subnet"
-path = "consensus/tecdsa/tecdsa_signature_from_other_subnet_test.rs"
-
-[[bin]]
-name = "ic-systest-tecdsa-signature-same-subnet"
-path = "consensus/tecdsa/tecdsa_signature_same_subnet_test.rs"
-
-[[bin]]
-name = "ic-systest-tecdsa-signature-timeout"
-path = "consensus/tecdsa/tecdsa_signature_timeout_test.rs"
-
-[[bin]]
-name = "ic-systest-tecdsa-two-signing-subnets"
-path = "consensus/tecdsa/tecdsa_two_signing_subnets_test.rs"
-
-[[bin]]
 name = "ic-systest-btc-get-balance"
 path = "financial_integrations/btc_get_balance_test.rs"
