+++ rs/tests/consensus/BUILD.bazel
@@ -25,6 +25,7 @@ system_test(
         "//rs/recovery",
         "//rs/registry/subnet_type",
         "//rs/tests",
+        "//rs/tests/consensus/tecdsa/utils",
         "//rs/tests/consensus/utils",
         "//rs/tests/driver:ic-system-test-driver",
         "//rs/types/base_types",
