+++ rs/tests/consensus/backup_manager_test.rs
@@ -35,6 +35,7 @@ use ic_consensus_system_test_utils::{
         wait_until_authentication_is_granted, AuthMean,
     },
 };
+use ic_consensus_threshold_sig_system_test_utils::run_chain_key_signature_test;
 use ic_management_canister_types::{
     EcdsaCurve, EcdsaKeyId, MasterPublicKeyId, SchnorrAlgorithm, SchnorrKeyId,
 };
@@ -51,7 +52,7 @@ use ic_system_test_driver::{
     util::{block_on, get_nns_node, MessageCanister, UniversalCanister},
 };
 use ic_tests::orchestrator::utils::{
-    subnet_recovery::{enable_chain_key_on_subnet, run_chain_key_signature_test},
+    subnet_recovery::enable_chain_key_on_subnet,
     upgrade::{
         assert_assigned_replica_version, bless_public_replica_version,
         deploy_guestos_to_all_subnet_nodes, get_assigned_replica_version, UpdateImageType,
