+++ rs/tests/consensus/tecdsa/Cargo.toml
@@ -11,6 +11,7 @@ anyhow = { workspace = true }
 canister-test = { path = "../../../rust_canisters/canister_test" }
 chrono = { workspace = true }
 futures = { workspace = true }
+ic-agent = { workspace = true }
 ic_consensus_system_test_utils = { path = "../utils" }
 ic_consensus_threshold_sig_system_test_utils = { path = "./utils" }
 ic-config = { path = "../../../config" }
@@ -21,9 +22,11 @@ ic-nns-constants = { path = "../../../nns/constants" }
 ic-nns-governance = { path = "../../../nns/governance" }
 ic-registry-subnet-features = { path = "../../../registry/subnet_features" }
 ic-registry-subnet-type = { path = "../../../registry/subnet_type" }
+ic-registry-nns-data-provider = { path = "../../../registry/nns_data_provider" }
 ic-sys = { path = "../../../sys" }
 ic-system-test-driver = { path = "../../driver" }
 ic-types = { path = "../../../types/types" }
+itertools = { workspace = true }
 registry-canister = { path = "../../../registry/canister" }
 reqwest = { workspace = true }
 rand = { workspace = true }
@@ -46,41 +49,41 @@ path = "tecdsa_add_nodes_test.rs"
 name = "ic-systest-tschnorr-message-sizes"
 path = "tschnorr_message_sizes_test.rs"
 
-# [[bin]]
-# name = "ic-systest-tecdsa-key-rotation"
-# path = "tecdsa_key_rotation_test.rs"
+[[bin]]
+name = "ic-systest-tecdsa-key-rotation"
+path = "tecdsa_key_rotation_test.rs"
 
 [[bin]]
 name = "ic-systest-tecdsa-remove-nodes"
 path = "tecdsa_remove_nodes_test.rs"
 
-# [[bin]]
-# name = "ic-systest-tecdsa-signature-life-cycle"
-# path = "tecdsa_signature_life_cycle_test.rs"
+[[bin]]
+name = "ic-systest-tecdsa-signature-life-cycle"
+path = "tecdsa_signature_life_cycle_test.rs"
 
-# [[bin]]
-# name = "ic-systest-tecdsa-signature-fails-without-cycles"
-# path = "tecdsa_signature_fails_without_cycles_test.rs"
+[[bin]]
+name = "ic-systest-tecdsa-signature-fails-without-cycles"
+path = "tecdsa_signature_fails_without_cycles_test.rs"
 
-# [[bin]]
-# name = "ic-systest-tecdsa-signature-from-nns-without-cycles"
-# path = "tecdsa_signature_from_nns_without_cycles_test.rs"
+[[bin]]
+name = "ic-systest-tecdsa-signature-from-nns-without-cycles"
+path = "tecdsa_signature_from_nns_without_cycles_test.rs"
 
-# [[bin]]
-# name = "ic-systest-tecdsa-signature-from-other-subnet"
-# path = "tecdsa_signature_from_other_subnet_test.rs"
+[[bin]]
+name = "ic-systest-tecdsa-signature-from-other-subnet"
+path = "tecdsa_signature_from_other_subnet_test.rs"
 
-# [[bin]]
-# name = "ic-systest-tecdsa-signature-same-subnet"
-# path = "tecdsa_signature_same_subnet_test.rs"
+[[bin]]
+name = "ic-systest-tecdsa-signature-same-subnet"
+path = "tecdsa_signature_same_subnet_test.rs"
 
-# [[bin]]
-# name = "ic-systest-tecdsa-signature-timeout"
-# path = "tecdsa_signature_timeout_test.rs"
+[[bin]]
+name = "ic-systest-tecdsa-signature-timeout"
+path = "tecdsa_signature_timeout_test.rs"
 
-# [[bin]]
-# name = "ic-systest-tecdsa-two-signing-subnets"
-# path = "tecdsa_two_signing_subnets_test.rs"
+[[bin]]
+name = "ic-systest-tecdsa-two-signing-subnets"
+path = "tecdsa_two_signing_subnets_test.rs"
 
 [[bin]]
 name = "ic-systest-tecdsa-complaint-test"
