+++ rs/tests/consensus/tecdsa/tecdsa_performance_test.rs
@@ -43,7 +43,9 @@ use futures::future::join_all;
 use ic_consensus_system_test_utils::{
     limit_tc_ssh_command, rw_message::install_nns_with_customizations_and_check_progress,
 };
-use ic_consensus_threshold_sig_system_test_utils::ChainSignatureRequest;
+use ic_consensus_threshold_sig_system_test_utils::{
+    run_chain_key_signature_test, ChainSignatureRequest,
+};
 use ic_management_canister_types::MasterPublicKeyId;
 use ic_registry_subnet_features::{ChainKeyConfig, KeyConfig};
 use ic_registry_subnet_type::SubnetType;
@@ -68,9 +70,7 @@ use ic_system_test_driver::systest;
 use ic_system_test_driver::util::{
     block_on, get_app_subnet_and_node, get_nns_node, MessageCanister,
 };
-use ic_tests::orchestrator::utils::subnet_recovery::{
-    enable_chain_key_signing_on_subnet, run_chain_key_signature_test,
-};
+use ic_tests::orchestrator::utils::subnet_recovery::enable_chain_key_signing_on_subnet;
 use ic_types::Height;
 use slog::{error, info};
 use std::fs::create_dir_all;
