+++ rs/tests/consensus/tecdsa/tecdsa_signature_from_nns_without_cycles_test.rs
@@ -1,19 +1,52 @@
-#[rustfmt::skip]
-
 use anyhow::Result;
 
-use ic_system_test_driver::driver::group::SystemTestGroup;
-use ic_system_test_driver::systest;
-use ic_tests::tecdsa::tecdsa_signature_test::{
-    config, test_threshold_ecdsa_signature_from_nns_without_cycles,
+use canister_test::Canister;
+use ic_consensus_threshold_sig_system_test_utils::{
+    enable_chain_key_signing, get_public_key_and_test_signature, make_key_ids_for_all_schemes,
+    setup,
 };
+use ic_nns_constants::GOVERNANCE_CANISTER_ID;
+use ic_registry_subnet_type::SubnetType;
+use ic_system_test_driver::{
+    driver::{
+        group::SystemTestGroup,
+        test_env::TestEnv,
+        test_env_api::{HasPublicApiUrl, HasTopologySnapshot, IcNodeContainer},
+    },
+    systest,
+    util::{block_on, runtime_from_url, MessageCanister},
+};
+
+/// Tests that a threshold signature request coming from the NNS succeeds even when
+/// there are no cycles sent with the request.
+fn test(env: TestEnv) {
+    let log = env.logger();
+    let topology = env.topology_snapshot();
+    let nns_subnet = topology.root_subnet();
+    let app_subnet = topology
+        .subnets()
+        .find(|s| s.subnet_type() == SubnetType::Application)
+        .unwrap();
+    let nns_node = nns_subnet.nodes().next().unwrap();
+    let nns_agent = nns_node.build_default_agent();
+    block_on(async move {
+        let nns = runtime_from_url(nns_node.get_public_url(), nns_node.effective_canister_id());
+        let governance = Canister::new(&nns, GOVERNANCE_CANISTER_ID);
+        let key_ids = make_key_ids_for_all_schemes();
+        enable_chain_key_signing(&governance, app_subnet.subnet_id, key_ids.clone(), &log).await;
+        let msg_can = MessageCanister::new(&nns_agent, nns_node.effective_canister_id()).await;
+        for key_id in &key_ids {
+            let _public_key = get_public_key_and_test_signature(key_id, &msg_can, true, &log)
+                .await
+                .unwrap();
+        }
+    });
+}
 
 fn main() -> Result<()> {
     SystemTestGroup::new()
-        .with_setup(config)
-        .add_test(systest!(
-            test_threshold_ecdsa_signature_from_nns_without_cycles
-        ))
+        .with_setup(setup)
+        .add_test(systest!(test))
         .execute_from_args()?;
     Ok(())
 }
