+++ rs/tests/src/orchestrator/upgrade_downgrade.rs
@@ -12,10 +12,7 @@ Success:: Upgrades work into both directions for all subnet types.
 
 end::catalog[] */
 
-use crate::orchestrator::utils::{
-    subnet_recovery::{enable_chain_key_signing_on_subnet, run_chain_key_signature_test},
-    upgrade::*,
-};
+use crate::orchestrator::utils::{subnet_recovery::enable_chain_key_signing_on_subnet, upgrade::*};
 use candid::Principal;
 use futures::future::join_all;
 use ic_agent::Agent;
@@ -24,7 +21,7 @@ use ic_consensus_system_test_utils::rw_message::{
     install_nns_and_check_progress, store_message,
 };
 use ic_consensus_threshold_sig_system_test_utils::{
-    make_key_ids_for_all_schemes, ChainSignatureRequest,
+    make_key_ids_for_all_schemes, run_chain_key_signature_test, ChainSignatureRequest,
 };
 use ic_management_canister_types::MasterPublicKeyId;
 use ic_registry_subnet_features::{ChainKeyConfig, KeyConfig, DEFAULT_ECDSA_MAX_QUEUE_SIZE};
