+++ rs/tests/src/orchestrator/utils/subnet_recovery.rs
@@ -15,8 +15,8 @@ use ic_consensus_system_test_utils::{
 };
 use ic_consensus_threshold_sig_system_test_utils::{
     add_chain_keys_with_timeout_and_rotation_period, create_new_subnet_with_keys,
-    empty_subnet_update, execute_update_subnet_proposal, get_public_key_with_retries,
-    get_signature_with_logger, verify_signature,
+    empty_subnet_update, execute_update_subnet_proposal, get_master_public_key,
+    get_signature_with_logger,
 };
 use ic_management_canister_types::MasterPublicKeyId;
 use ic_nns_constants::GOVERNANCE_CANISTER_ID;
@@ -447,47 +447,4 @@ pub(crate) fn disable_chain_key_on_subnet(
         )
         .expect("Failed to detect disabled signing.");
     }
-}
-
-/// Get the threshold public key of the given canister
-pub(crate) fn get_master_public_key(
-    canister: &MessageCanister,
-    key_id: &MasterPublicKeyId,
-    logger: &Logger,
-) -> Vec<u8> {
-    info!(
-        logger,
-        "Getting threshold public key for key id: {}.", key_id
-    );
-    let public_key = block_on(get_public_key_with_retries(key_id, canister, logger, 100)).unwrap();
-    info!(logger, "Got public key {:?}", public_key);
-    public_key
-}
-
-/// The signature test consists of getting the given canister's Chain key, comparing it to the existing key
-/// to ensure it hasn't changed, sending a sign request, and verifying the signature
-pub fn run_chain_key_signature_test(
-    canister: &MessageCanister,
-    logger: &Logger,
-    key_id: &MasterPublicKeyId,
-    existing_key: Vec<u8>,
-) {
-    info!(logger, "Run through Chain key signature test.");
-    let message_hash = vec![0xabu8; 32];
-    block_on(async {
-        let public_key = get_public_key_with_retries(key_id, canister, logger, 100)
-            .await
-            .unwrap();
-        assert_eq!(existing_key, public_key);
-        let signature = get_signature_with_logger(
-            message_hash.clone(),
-            ECDSA_SIGNATURE_FEE,
-            key_id,
-            canister,
-            logger,
-        )
-        .await
-        .unwrap();
-        verify_signature(key_id, &message_hash, &public_key, &signature);
-    });
 }
