--- rs/tests/src/tecdsa/mod.rs
@@ -1,2 +1,0 @@
-pub mod tecdsa_signature_test;
-pub mod tecdsa_two_signing_subnets_test;
